Fasal-Mitra/DeepLearningproject/api/requirements.txt
# Web Framework and Server
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
starlette==0.27.0
pydantic==2.4.2
typing-extensions==4.8.0

# AI/ML Dependencies
tensorflow==2.14.0
numpy==1.24.3
Pillow==10.0.0
scikit-learn==1.3.2
opencv-python-headless==4.8.1.78

# Policy Scraping and Processing
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
html5lib==1.1

# Utilities
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
aiofiles==23.2.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# CORS and Security
cors==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Monitoring and Logging
prometheus-client==0.17.1
structlog==23.1.0

# Development Tools
black==23.9.1
flake8==6.1.0
pytest==7.4.2